Fix stats recalculation issue
